+++ packages/catalog-schemas/src/index.ts
@@ -2,6 +2,7 @@ export * from './constants';
 export * from './types';
 export * from './components/simple-table';
 export * from './components/flex-table';
+export * from './components/html-table';
 export * from './components/latex-formula';
 export * from './components/choice-card';
 export * from './components/form-card';